requirements.txt
streamlit
groq
python-dotenv
plotly
pandas
